HomeWork
after comments
